// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
};
// ...
export type Competition = {
  __typename?: 'Competition';
  area: Area;
  id: Scalars['ID']['output'];
  image: Image;
  name: Scalars['String']['output'];
};
// ...
export type Image = {
  __typename?: 'Image';
  url: Scalars['String']['output'];
};

export type Injury = {
  __typename?: 'Injury';
  description?: Maybe<Scalars['String']['output']>;
  endDate?: Maybe<Scalars['DateTime']['output']>;
  person?: Maybe<Person>;
  shirtNumber?: Maybe<Scalars['Int']['output']>;
};

export type LineupPlayer = {
  __typename?: 'LineupPlayer';
  events?: Maybe<Array<Maybe<DataEvent>>>;
  isCaptain?: Maybe<Scalars['Boolean']['output']>;
  person?: Maybe<Person>;
  pitchPosition?: Maybe<PitchPosition>;
  score?: Maybe<Scalars['Int']['output']>;
  shirtNumber?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Manager = {
  __typename?: 'Manager';
  id?: Maybe<Scalars['ID']['output']>;
  image?: Maybe<ManagerImage>;
  name?: Maybe<Scalars['String']['output']>;
};
// ...
export type Nationality = {
  __typename?: 'Nationality';
  image?: Maybe<Image>;
  name?: Maybe<Scalars['String']['output']>;
};
// ...
export type PlayerDetails = {
  __typename?: 'PlayerDetails';
  age?: Maybe<Scalars['Int']['output']>;
  dateOfBirth?: Maybe<Scalars['DateTime']['output']>;
  firstName?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  image?: Maybe<Image>;
  lastName?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  nationality?: Maybe<Nationality>;
  position?: Maybe<Scalars['String']['output']>;
  shirtNumber?: Maybe<Scalars['Int']['output']>;
  stats?: Maybe<Array<Maybe<PlayerStats>>>;
  team?: Maybe<Team>;
};

export type PlayerStatTeam = {
  __typename?: 'PlayerStatTeam';
  id: Scalars['ID']['output'];
  image?: Maybe<Image>;
};

export type PlayerStats = {
  __typename?: 'PlayerStats';
  competition?: Maybe<Competition>;
  season?: Maybe<Season>;
  stats?: Maybe<StatItems>;
  team?: Maybe<PlayerStatTeam>;
};
// ...
export type Season = {
  __typename?: 'Season';
  active?: Maybe<Scalars['Boolean']['output']>;
  id?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
};
// ...
export type StatItems = {
  __typename?: 'StatItems';
  appearances?: Maybe<Scalars['Int']['output']>;
  assists?: Maybe<Scalars['Int']['output']>;
  blockedShots?: Maybe<Scalars['Int']['output']>;
  cleanSheets?: Maybe<Scalars['Int']['output']>;
  clearances?: Maybe<Scalars['Int']['output']>;
  corners?: Maybe<Scalars['Int']['output']>;
  crosses?: Maybe<Scalars['Int']['output']>;
  foulsCommited?: Maybe<Scalars['Int']['output']>;
  foulsSuffered?: Maybe<Scalars['Int']['output']>;
  freekickGoals?: Maybe<Scalars['Int']['output']>;
  goals?: Maybe<Scalars['Int']['output']>;
  goalsConceded?: Maybe<Scalars['Int']['output']>;
  goalsOutsideBox?: Maybe<Scalars['Int']['output']>;
  hitWoodwork?: Maybe<Scalars['Int']['output']>;
  minutesPerGoal?: Maybe<Scalars['Int']['output']>;
  minutesPlayed?: Maybe<Scalars['Int']['output']>;
  offsides?: Maybe<Scalars['Int']['output']>;
  ownGoals?: Maybe<Scalars['Int']['output']>;
  penaltiesMissed?: Maybe<Scalars['Int']['output']>;
  penaltyGoals?: Maybe<Scalars['Int']['output']>;
  penaltySaves?: Maybe<Scalars['Int']['output']>;
  redCards?: Maybe<Scalars['Int']['output']>;
  saves?: Maybe<Scalars['Int']['output']>;
  shotsOffTarget?: Maybe<Scalars['Int']['output']>;
  shotsOnTarget?: Maybe<Scalars['Int']['output']>;
  startingEleven?: Maybe<Scalars['Int']['output']>;
  successfulCrosses?: Maybe<Scalars['Int']['output']>;
  tackles?: Maybe<Scalars['Int']['output']>;
  yellowCards?: Maybe<Scalars['Int']['output']>;
};

export type Stats = {
  __typename?: 'Stats';
  gamesBothTeamsScored?: Maybe<Scalars['Int']['output']>;
  gamesOverTwoAndHalfGoals?: Maybe<Scalars['Int']['output']>;
  gamesTotal?: Maybe<Scalars['Int']['output']>;
  goalsConceded?: Maybe<Scalars['Int']['output']>;
  goalsScored?: Maybe<Scalars['Int']['output']>;
};

export type Substitute = {
  __typename?: 'Substitute';
  events?: Maybe<Array<Maybe<DataEvent>>>;
  person?: Maybe<Person>;
  score?: Maybe<Scalars['Int']['output']>;
  shirtNumber?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Team = {
  __typename?: 'Team';
  code: Scalars['String']['output'];
  formation?: Maybe<Scalars['String']['output']>;
  full: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  image: Image;
  injuries?: Maybe<Array<Maybe<Injury>>>;
  kit?: Maybe<Scalars['String']['output']>;
  lineup?: Maybe<Array<Maybe<LineupPlayer>>>;
  long: Scalars['String']['output'];
  manager?: Maybe<Manager>;
  name: Scalars['String']['output'];
  short: Scalars['String']['output'];
  substitutes?: Maybe<Array<Maybe<Substitute>>>;
  suspensions?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
};